# api-contracts/openapi/v1/export.yaml
openapi: 3.1.0
info:
  title: Skeldir Export API
  version: 2.0.0
  description: |
    Data export functionality for CSV, JSON, and Excel formats.
    Supports scheduled exports and custom date ranges.
  contact:
    name: Skeldir Engineering
    email: engineering@skeldir.com
    url: https://skeldir.com

servers:
  - url: http://localhost:4013
    description: Prism Mock Server (Export)

tags:
  - name: Export
    description: Data export functionality for CSV, JSON, and Excel formats

security:
  - accessBearerAuth: ["viewer"]

paths:
  /api/export/revenue:
    get:
      summary: Export revenue data
      description: Export revenue and attribution data in specified format
      operationId: exportRevenue
      tags:
        - Export
      security:
        - accessBearerAuth: ["viewer"]
      parameters:
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [csv, json, xlsx]
            default: csv
          description: Export file format
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date for export range
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date for export range
        - name: channels
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          description: Filter by specific channels
      responses:
        '200':
          description: Export file
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
            Content-Disposition:
              schema:
                type: string
              description: Attachment filename
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/ExportData'
              example:
                generated_at: '2025-11-26T14:30:00Z'
                date_range:
                  start: '2025-11-01'
                  end: '2025-11-26'
                data:
                  - date: '2025-11-25'
                    channel: Meta
                    revenue: 15230.50
                    conversions: 127
                    confidence: 0.92
                  - date: '2025-11-25'
                    channel: Google
                    revenue: 12450.00
                    conversions: 98
                    confidence: 0.89
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'
        '403':
          $ref: './_common/base.yaml#/components/responses/ForbiddenError'
        '400':
          $ref: './_common/base.yaml#/components/responses/ValidationError'

  /api/export/csv:
    get:
      summary: Export data as CSV
      description: Export attribution data in CSV format
      operationId: exportCSV
      tags:
        - Export
      security:
        - accessBearerAuth: ["viewer"]
      parameters:
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
      responses:
        '200':
          description: CSV file
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="skeldir-export-2025-11-26.csv"
          content:
            text/csv:
              schema:
                type: string
              example: |
                date,channel,revenue,conversions,confidence
                2025-11-25,Meta,15230.50,127,0.92
                2025-11-25,Google,12450.00,98,0.89
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'
        '403':
          $ref: './_common/base.yaml#/components/responses/ForbiddenError'

  /api/export/json:
    get:
      summary: Export data as JSON
      description: Export attribution data in JSON format
      operationId: exportJSON
      tags:
        - Export
      security:
        - accessBearerAuth: ["viewer"]
      parameters:
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
      responses:
        '200':
          description: JSON data
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportData'
              example:
                generated_at: '2025-11-26T14:30:00Z'
                date_range:
                  start: '2025-11-01'
                  end: '2025-11-26'
                data:
                  - date: '2025-11-25'
                    channel: Meta
                    revenue: 15230.50
                    conversions: 127
                    confidence: 0.92
                  - date: '2025-11-25'
                    channel: Google
                    revenue: 12450.00
                    conversions: 98
                    confidence: 0.89
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'
        '403':
          $ref: './_common/base.yaml#/components/responses/ForbiddenError'

  /api/export/excel:
    get:
      summary: Export data as Excel
      description: Export attribution data in Excel (XLSX) format
      operationId: exportExcel
      tags:
        - Export
      security:
        - accessBearerAuth: ["viewer"]
      parameters:
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
      responses:
        '200':
          description: Excel file
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="skeldir-export-2025-11-26.xlsx"
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'
        '403':
          $ref: './_common/base.yaml#/components/responses/ForbiddenError'

components:
  schemas:
    ExportData:
      type: object
      required:
        - generated_at
        - date_range
        - data
      properties:
        generated_at:
          type: string
          format: date-time
          description: When the export was generated
        date_range:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              channel:
                type: string
              revenue:
                type: number
                format: float
              conversions:
                type: integer
              confidence:
                type: number
                format: float
        summary:
          type: object
          properties:
            total_revenue:
              type: number
              format: double
              minimum: 0
            total_conversions:
              type: integer
            average_confidence:
              type: number
              format: double
              minimum: 0
              maximum: 1

  securitySchemes:
    accessBearerAuth:
      $ref: './_common/base.yaml#/components/securitySchemes/accessBearerAuth'